Convert OlmDevice to Typescript
